// src/types.ts
export interface HistoryMeta {
  type?: string;
  time?: number;
  label?: string;
}

export interface HistoryState<S> {
  past: S[];
  present: S;
  future: S[];
  pastMeta: HistoryMeta[];
  presentMeta: HistoryMeta;
  futureMeta: HistoryMeta[];
}